#63-fix: 게시글 삭제를 별명이 일치할 때로 바꾸었다..

+++ geulkkoli/src/main/java/com/geulkkoli/domain/post/service/PostService.java
@@ -91,7 +91,12 @@ public class PostService {
         postRepository.save(post);
     }
     @Transactional
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId , String nickName) {
+        Post post = postRepository.findById(postId)
+                .orElseThrow(() -> new NoSuchElementException("No post found id matches:" + postId));
+        if (!post.getNickName().equals(nickName)) {
+            throw new NoSuchElementException("No post found id matches:" + postId);
+        }
         postRepository.deleteById(postId);
     }
